Move FastCopyable
