+++ platform-sdk/base-utility/src/main/java/org/hiero/base/FastCopyable.java
@@ -1,10 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common;
+package org.hiero.base;
 
 import com.swirlds.base.state.Mutable;
-import org.hiero.base.Copyable;
-import org.hiero.base.Releasable;
-import org.hiero.base.crypto.Hashable;
 
 /**
  * An interface for classes that can be copied and serialized in a way specific to the Swirlds platform. If a class
@@ -33,7 +30,7 @@ public interface FastCopyable extends Copyable, Mutable, Releasable {
      * Either behavior is fine, but each implementation should document which behavior it has chosen.
      * By the default, the first implementation is assumed.
      *
-     * If a FastCopyable object extends {@link Hashable} then under no circumstances should
+     * If a FastCopyable object extends {@code Hashable} then under no circumstances should
      * the hash be copied by this method.
      *
      * It is strongly suggested that each implementing class override the return type of this method to its self
