+++ platform-sdk/consensus-state/src/main/java/module-info.java
@@ -10,13 +10,13 @@ module org.hiero.consensus.state {
     requires transitive com.hedera.node.hapi;
     requires transitive com.hedera.pbj.runtime;
     requires transitive com.swirlds.base;
-    requires transitive com.swirlds.common;
     requires transitive com.swirlds.config.api;
     requires transitive com.swirlds.metrics.api;
     requires transitive com.swirlds.state.impl;
     requires transitive org.hiero.base.crypto;
     requires transitive org.hiero.base.utility;
     requires transitive org.hiero.consensus.model;
+    requires com.swirlds.common;
     requires com.swirlds.logging;
     requires com.swirlds.state.api;
     requires com.swirlds.virtualmap;
